add Android library with activity for telling jokes and wire it up to MainActvity
Satisfies these steps from the project instructions:

Step 2: Create an Android Library
------------------------------------
Create an Android Library containing an Activity that will display a joke passed to it as an intent extra.
Wire up project dependencies so that the button can now pass the joke from the Java Library to the Android Library.
For review on how to create an Android library, check out demo 4.03. For a refresher on intent extras,
check out this documentation.